fix: 스트레스 계산 ë¡ 알고리즘 롤백

+++ backend/fast-api/models/request_models.py
@@ -2,7 +2,7 @@ from pydantic import BaseModel
 
 
 class SingleStressRequest(BaseModel):
-    hr_data: list[int]
+    hr_data: list[float]
 
 
 class StressOverviewRequest(BaseModel):
